+++ AndibagProject/app/src/main/java/com/example/andibagproject/data/gallery/reply/ReplyRepository.kt
@@ -9,17 +9,17 @@ import io.reactivex.rxjava3.schedulers.Schedulers
 import retrofit2.Response
 
 class ReplyRepository {
-    fun writeReply(content: String, id: Int) : @NonNull Single<Response<Unit>> =
+    fun writeReply(content: String, id: Long) : @NonNull Single<Response<Unit>> =
         replyAPI.writeReply(ACCESS_TOKEN, content, id)
             .observeOn(AndroidSchedulers.mainThread())
             .subscribeOn(Schedulers.io())
 
-    fun deleteReply(id: Int) : @NonNull Single<Response<Unit>> =
+    fun deleteReply(id: Long) : @NonNull Single<Response<Unit>> =
         replyAPI.deleteReply(ACCESS_TOKEN, id)
             .observeOn(AndroidSchedulers.mainThread())
             .subscribeOn(Schedulers.io())
 
-    fun patchReply(content: String, id: Int) : @NonNull Single<Response<Unit>> =
+    fun patchReply(content: String, id: Long) : @NonNull Single<Response<Unit>> =
         replyAPI.patchReply(ACCESS_TOKEN, content, id)
             .observeOn(AndroidSchedulers.mainThread())
             .subscribeOn(Schedulers.io())
